[feat, fix]Member/Bookmarks 코드 update, oauth2 인증 및 세션 설정,  컴포넌트 생성
1.  Member와 Bookmarks 관련된 모든 코드에서
- ItemResponseDTO 받는 메소드들 -> 전부 Item을 받도록 수정
2. oauth2 인증, 세션 관련 설정 추가:
- application-oauth.properties 생성 및 등록
- application-local properties에 세션이 올라갈 DB 지정
3. BookmarksService:
- 북마크 추가/삭제 메소드(toggle) 리팩토링
- 북마크 단건 조회, 다중 조회 메소드(getBookmark, getMyBookmarkList) 완성
4. 회원가입, 로그인, 아이디 찾기/비밀번호 재설정, 메인 컴포넌트 생성
- components 폴더 생성
- Signup.js, Signup_site.js / Login.js, FindId.js, FindId_result.js, ChangePw.js, ChangePw_result.js / main.js, Header.js 생성

+++ threeh/src/main/java/com/cmyk/threeh/service/MemberService.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.cmyk.threeh.domain.Member;
 import com.cmyk.threeh.dto.MemberDTO;
@@ -18,6 +19,7 @@ import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class MemberService {
     
     private final MemberRepository memberRepository;
@@ -26,6 +28,7 @@ public class MemberService {
 
 	//<계정 관리>
     //회원가입
+	@Transactional
     public boolean create(SignupUpdateForm form) {
 		
 		//회원이 입력한 정보
@@ -41,8 +44,8 @@ public class MemberService {
 		//중복체크
 		boolean isExistId = memberRepository.existsById(id);
 		boolean isExistEmail = memberRepository.existsByEmail(email);
-		boolean isExistPhone = memberRepository.existsById(phone);
-		boolean isExistRegNo = memberRepository.existsById(regNo);
+		boolean isExistPhone = memberRepository.existsByPhone(phone);
+		boolean isExistRegNo = memberRepository.existsByRegNo(regNo);
 
 		if (isExistId) throw new CustomException(ErrorCode.MEMBER_FOUND);
 		if (isExistEmail) throw new CustomException(ErrorCode.EMAIL_IS_EXIST);
@@ -97,6 +100,7 @@ public class MemberService {
 
 
 	//회원정보 수정
+	@Transactional
 	public boolean update(Member member, SignupUpdateForm form) {
 
 		member.setPassword(passwordEncoder.encode(form.getPassword1()));
@@ -113,9 +117,14 @@ public class MemberService {
 
 
 	//회원 탈퇴
+	@Transactional
 	public boolean delete(String id) {
 		
-		Member member = getUser(id);
+		Optional<Member> op = memberRepository.findById(id);
+		if(!op.isPresent()) {
+			throw new CustomException(ErrorCode.MEMBER_NOT_FOUND);
+		}
+		Member member = op.get();
 
 		memberRepository.delete(member);
 		return true;
@@ -126,7 +135,12 @@ public class MemberService {
 	//회원 role 가져오기
 	public String checkMemberRole(String id) {
 
-		Member member = getUser(id);
+		Optional<Member> op = memberRepository.findById(id);
+		if(!op.isPresent()) {
+			throw new CustomException(ErrorCode.MEMBER_NOT_FOUND);
+		}
+		Member member = op.get();
+
         String UserRole = member.getRole().getKey();
 		String result = null;
 
@@ -191,15 +205,19 @@ public class MemberService {
 
 
 	//비밀번호 재설정
+	@Transactional
 	public boolean changeUserPassword(String id, String oldPassword, String newPassword) {
 		
-		Member member = null;
+		Optional<Member> op = memberRepository.findById(id);
+		if(!op.isPresent()) {
+			throw new CustomException(ErrorCode.MEMBER_NOT_FOUND);
+		}
+		Member member = op.get();
 
 		//1. 로그인 상태인지 검사한다.
 		// (세션에 유저 정보가 올라가 있으면, 로그인 상태)
 
 		//2. 회원가입이 되어있는 상태인지 검사한다. (세션 유저 정보에서 가져옴)
-		member = getUser(id);
 		if(member==null) {
 			throw new CustomException(ErrorCode.MEMBER_NOT_FOUND);
 		}
